final_files/prediction-api/requirements.txt
setuptools
flask
pandas
scikit-learn==1.3.2
pytest
xgboost
google-cloud-storage